ai_job_backend/requirements.txt
# After pip install, run: playwright install  (browser binaries).
# On Linux/WSL also run: sudo playwright install-deps  (system libs for Chromium).
fastapi==0.115.6
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.55.3
azure-ai-inference==1.0.0b4
playwright==1.49.0
beautifulsoup4==4.12.3
supabase==2.10.0
PyJWT[crypto]==2.10.1
pypdf==5.1.0
python-docx==1.1.2